Minor update to example, changed type from uint64_t to int

+++ examples/billion_integers/billion_integers.cpp
@@ -4,7 +4,7 @@ int main()
 {
   binary_log::binary_log log("log.out");
 
-  for (uint64_t i = 0; i < 1E9; ++i) {
+  for (int i = 0; i < 1E9; ++i) {
     BINARY_LOG(log, "Hello logger, msg number: {}", i);
   }
 }
